remove the consumer callbacks and make an internal message queue
this simplifies the types and make it easier to implement with futures.
When using in a lower level context, poll the consumers directly for
new messages after handling frames.

it might also be possible to stream the message content by reading
the content of the `current_message` field of `Consumer`