futures: "force" users to declare queue before consuming
consuming won't work if the queue hasn't been declared in the same channel first.
introduce a new type returned by queue_declare and taken instead of the queue name in the basic_consume api.
This doesn't actually enforce anything but it's a good start to guide the users in the right direction

Signed-off-by: Marc-Antoine Perennou <Marc-Antoine@Perennou.com>