Introduce dedicated error types to the lapin-futures crate
This change replaces all occurrences of io::Error by the newly created
Error type, the only exception being for AMQPCodec. This change allows
future-proofing the error that can be returned by the crate. For example
by changing the internals of the Error type, it is now possible to
provide users with a way of knowing what kind of error occurred without
having to resort to parsing error messages.

The only exception to this change concerns AMQPCodec: a new Error type
was created just for it because tokio-codec's Encoder and Decoder traits
impose a From<std::io::Error> bound of the error type. Because I didn't
want to introduce a way to create an Error instance from outside the
crate, I isolated the possible errors from this section into its own
type.