IoLoop: rewrite should_continue condition
This makes it way more readable.
Also, we now make sure that we properly continue writing to the TCP
stream if w have pending writes and it's not in error.
This enables us not to leave promises behind and properly answering to a
rabbitmq server's connection.close with a connection.close-ok.

Fixes #413

Properly fixes #409

Signed-off-by: Marc-Antoine Perennou <Marc-Antoine@Perennou.com>