avoid deadlock in server protocol error codepath
When the server sends us a content header frame for an invalid protocol class id, we close the connection and emit an error from a callback.
The connection status lock is held when running this callback, so we cannot call external methods that try and acquire the lock.
Let's use our internal RPC to avoid this by delaying the call to after we've released the lock, in the RPC thread.

Fixes #429