internal_rpc: track channel killswitch
When reconnecting a channel, we activate the current killswitch and replace it with a new one.
Set an internal state in the internal rpc handler.
When we queue channel actions, if we reconnect, we queue the update of the killswitch and all the requests before this one that are pending will get discarded because the current (old) killswitch has been activated