Fix connection with tokio
Recent changes moved the reactor register in the io_loop.
The io_loop thread is not managed by an executor, and tokio doesn't like
this.
This could and should be fixed in tokio-reactor-trait but requires a
breaking change there.
In the meantime, let's just move that part to the internal RPC task
which runs inside the executor.
While at it, make sure the tokio example runs properly by integrating in
the tests.

Fixes #436