avoid using two promises for synchronous methods
Intead of having one promise for the method sending, and another for the
reply, let's use only one. We only care about errors for the sending
anyways, so we go through the Cancelable trait for this, using the same
resolver.
We still need to actually resolve the promise on sending for
asynchronous methods though, such as publish and ack, so let's special
case these by sending the actual resolver to the io loop.