connection_status: fix deadlock when poisoning connection (#446)
When poisoning the connection status during an early network failure, we would end up dropping the connection held by the connection resolver while the ConnectionStatus is still locked.
Make sure we always release the lock before dropping the connection so that the ConnectionCloser doesn't lock us when accessing the ConnectionStatus.

Fixes #447